requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohappyeyeballs=2.4.0=py311haa95532_0
aiohttp=3.10.5=py311h827c3e9_0
aiosignal=1.2.0=pyhd3eb1b0_0
annotated-types=0.6.0=py311haa95532_0
anyio=4.2.0=py311haa95532_0
arrow-cpp=16.1.0=h7cd61ee_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.1.0=py311haa95532_0
aws-c-auth=0.6.19=h2bbff1b_0
aws-c-cal=0.5.20=h2bbff1b_0
aws-c-common=0.8.5=h2bbff1b_0
aws-c-compression=0.2.16=h2bbff1b_0
aws-c-event-stream=0.2.15=hd77b12b_0
aws-c-http=0.6.25=h2bbff1b_0
aws-c-io=0.13.10=h2bbff1b_0
aws-c-mqtt=0.7.13=h2bbff1b_0
aws-c-s3=0.1.51=h2bbff1b_0
aws-c-sdkutils=0.1.6=h2bbff1b_0
aws-checksums=0.1.13=h2bbff1b_0
aws-crt-cpp=0.18.16=hd77b12b_0
aws-sdk-cpp=1.10.55=hd77b12b_0
blas=1.0=mkl
boost-cpp=1.82.0=h59b6b97_2
bottleneck=1.3.7=py311hd7041d2_0
brotli-python=1.0.9=py311hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2024.8.30=h56e8100_0
certifi=2024.8.30=py311haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py311haa95532_0
colorama=0.4.6=py311haa95532_0
comm=0.2.2=pyhd8ed1ab_0
cuda-cccl=12.4.127=0
cuda-cudart=12.4.127=0
cuda-cudart-dev=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.0=0
cuda-libraries-dev=12.4.0=0
cuda-nvrtc=12.4.127=0
cuda-nvrtc-dev=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.4.127=0
cuda-opencl-dev=12.4.127=0
cuda-profiler-api=12.4.127=0
cuda-runtime=12.4.0=0
cuda-version=11.8=hcce14f8_3
datasets=2.19.1=py311haa95532_0
debugpy=1.8.5=py311hda3d55a_1
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.8=py311haa95532_0
distro=1.9.0=py311haa95532_0
einops=0.8.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
faiss-cpu=1.8.0=py3.11_h2272212_0_cpu
filelock=3.13.1=py311haa95532_0
flash-attn=2.6.3=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.0=py311h2bbff1b_0
fsspec=2024.3.1=py311haa95532_0
gflags=2.2.2=hd77b12b_1
glog=0.5.0=hd77b12b_1
gmpy2=2.1.2=py311h7f96b67_0
h11=0.14.0=py311haa95532_0
httpcore=1.0.2=py311haa95532_0
httpx=0.27.0=py311haa95532_0
huggingface_hub=0.24.6=py311haa95532_0
icc_rt=2022.1.0=h6049295_2
idna=3.7=py311haa95532_0
importlib-metadata=7.0.1=py311haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.5=pyh4bbf305_0
ipython=8.27.0=pyh7428d3b_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=py311haa95532_0
jpeg=9e=h827c3e9_3
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=py311h1ea47a8_0
krb5=1.21.3=hdf4eb48_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libabseil=20240116.2=cxx17_h5da7b33_0
libboost=1.82.0=h3399ecb_2
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libcublas=12.4.2.65=0
libcublas-dev=12.4.2.65=0
libcufft=11.2.0.44=0
libcufft-dev=11.2.0.44=0
libcurand=10.3.5.147=0
libcurand-dev=10.3.5.147=0
libcurl=8.9.1=h0416ee5_0
libcusolver=11.6.0.99=0
libcusolver-dev=11.6.0.99=0
libcusparse=12.3.0.142=0
libcusparse-dev=12.3.0.142=0
libdeflate=1.17=h2bbff1b_1
libevent=2.1.12=h56d1f94_1
libfaiss=1.8.0=h2e52968_0_cpu
libffi=3.4.4=hd77b12b_1
libgrpc=1.62.2=hf25190f_0
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.2.5.2=0
libnpp-dev=12.2.5.2=0
libnvfatbin=12.4.127=0
libnvfatbin-dev=12.4.127=0
libnvjitlink=12.4.99=0
libnvjitlink-dev=12.4.99=0
libnvjpeg=12.3.1.89=0
libnvjpeg-dev=12.3.1.89=0
libpng=1.6.39=h8cc25b3_0
libprotobuf=4.25.3=hf2fb9eb_0
libsentencepiece=0.2.0=h2f31864_2
libsodium=1.0.20=hc70643c_0
libssh2=1.11.0=h291bd65_0
libthrift=0.15.0=h4364b78_2
libtiff=4.5.1=hd77b12b_0
libuv=1.48.0=h827c3e9_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.3=py311h2bbff1b_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.10=py311h827c3e9_0
mkl_random=1.2.7=py311hea22821_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py311haa95532_0
multidict=6.0.4=py311h2bbff1b_0
multiprocess=0.70.15=py311haa95532_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.3=py311haa95532_0
ninja=1.10.2=haa95532_5
ninja-base=1.10.2=h6d14046_5
numexpr=2.8.7=py311h1fcbade_0
numpy=1.26.4=py311hdab7c0b_0
numpy-base=1.26.4=py311hd01c5d8_0
openai=1.37.0=py311haa95532_0
openjpeg=2.5.2=hae555c5_0
openssl=3.3.2=h2466b09_0
orc=2.0.1=hd8d391b_0
packaging=24.1=py311haa95532_0
pandas=2.2.2=py311hea22821_0
parso=0.8.4=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=py311h827c3e9_0
pip=24.2=py311haa95532_0
platformdirs=4.3.6=pyhd8ed1ab_0
prompt-toolkit=3.0.47=pyha770c72_0
protobuf=4.25.3=py311h958608f_0
psutil=6.0.0=py311he736701_1
pure_eval=0.2.3=pyhd8ed1ab_0
pyarrow=16.1.0=py311hea22821_0
pybind11-abi=5=hd3eb1b0_0
pydantic=2.8.2=py311haa95532_0
pydantic-core=2.20.1=py311hefb1915_0
pygments=2.18.0=pyhd8ed1ab_0
pysocks=1.7.1=py311haa95532_0
python=3.11.9=he1021f5_0
python-dateutil=2.9.0post0=py311haa95532_2
python-dotenv=0.21.0=py311haa95532_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py311h2bbff1b_1
python_abi=3.11=2_cp311
pytorch=2.4.1=py3.11_cuda12.4_cudnn9_0
pytorch-cuda=12.4=h3fd98bf_6
pytorch-mutex=1.0=cuda
pytz=2024.1=py311haa95532_0
pywin32=306=py311h12c1d0e_2
pyyaml=6.0.1=py311h2bbff1b_0
pyzmq=26.2.0=py311h484c95c_2
re2=2022.04.01=hd77b12b_0
regex=2024.9.11=py311h827c3e9_0
requests=2.32.3=py311haa95532_0
safetensors=0.4.4=py311h3ef4675_0
scipy=1.13.1=py311h9f229c6_0
sentencepiece=0.2.0=haa95532_2
sentencepiece-python=0.2.0=py311h2f31864_2
sentencepiece-spm=0.2.0=h2f31864_2
setuptools=75.1.0=py311haa95532_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.2.1=hcdb6601_0
sniffio=1.3.0=py311haa95532_0
sqlite=3.45.3=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.13.2=py311haa95532_0
tbb=2021.8.0=h59b6b97_0
tk=8.6.14=h0416ee5_0
tokenizers=0.19.1=py311h2da67c9_0
torchaudio=2.4.1=pypi_0
torchvision=0.19.1=pypi_0
tornado=6.4.1=py311he736701_1
tqdm=4.66.5=py311h746a85d_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.44.1=py311haa95532_0
typing-extensions=4.11.0=py311haa95532_0
typing_extensions=4.11.0=py311haa95532_0
tzdata=2024a=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.2.2=py311haa95532_0
utf8proc=2.6.1=h2bbff1b_1
vc=14.40=h2eaa2aa_1
vc14_runtime=14.40.33810=ha82c5b3_21
vs2015_runtime=14.40.33810=h3bf8584_21
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.44.0=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
xxhash=0.8.0=h2bbff1b_3
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.11.0=py311h827c3e9_0
zeromq=4.3.5=he1f189c_5
zipp=3.17.0=py311haa95532_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.5=hd43e919_2
